Step 3 of 4 for updating runtime C++ files for compliant naming, namespaces, and warnings

Closes #2679

See merge request opencpi/opencpi!762

+++ runtime/util/tcp/src/UtilTcpServer.cc
@@ -71,7 +71,7 @@ OCPI::Util::Tcp::Server::accept (unsigned long timeout)
     throw std::string ("not bound");
   }
 
-  if (!m_socket.wait (timeout)) {
+  if (!m_socket.wait((long)timeout)) {
     return 0;
   }
